[+] Fix update function accidentally made private
[+] Fix update function accidentally made private.\n
[+] State management of ViewModel moved to seperate mixin.
