+++ example/lib/pages/general/home/home_main_child/viewmodel/home_main_child_viewmodel.dart
@@ -3,7 +3,7 @@ import 'package:liman/product/extensions/extensions.dart';
 
 final class HomeMainChildViewModel extends BaseViewModel<HomeMainChildViewModel> {
   final count = 0.observable;
-  final weight = 0.5.observable;
+  final weight = 80.observable;
   int age = 26;
 
   void incrementCounter() {
